Refactor build views
